pyproject.toml
[project]
name = "rag-poc"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "langchain>=0.0.267",
    "langchain-community>=0.0.9",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.0.2",
    "langchain-qdrant>=0.0.1",
    "llvmlite>=0.40.0",
    "numpy>=1.24.0",
    "openai>=1.3.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "qdrant-client>=1.5.4",
    "scikit-learn>=1.3.0",
    "tiktoken>=0.5.1",
    "umap-learn>=0.5.4",
    "unstructured>=0.10.0",
    "unstructured-inference>=0.7.0",
    "uvicorn>=0.23.2",
]
